webpack

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
     const videos = await Video.find({}).sort({createdAt: "desc"});
@@ -6,7 +7,7 @@ export const home = async (req, res) => {
 }
 export const watch = async (req, res) => {
     const id = req.params.id;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if (!video){
         return res.render("404", {pageTitle: "Video not found."});
     }
@@ -15,19 +16,27 @@ export const watch = async (req, res) => {
 
 export const getedit = async (req, res) => {
     const id = req.params.id;
+    const {user:{_id}} = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     return res.render("edit", {pageTitle: `Editing ${video.title}`, video});
 }
 
 export const postedit = async (req, res) => {
     const id = req.params.id;
+    const {user:{_id}} = req.session;
     const {title, description, hashtags} = req.body;
-    const video = await Video.exists({_id: id});
+    const video = await Video.findById({id});
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id,{
         title,
@@ -41,15 +50,20 @@ export const getUpload = (req, res) => {
     return res.render("upload", {pageTitle:"Upload Video"});
 };
 export const postUpload = async (req, res) => {
+    const {user: {_id}} = req.session;
     const file = req.file;
     const {title, description, hashtags} = req.body;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             fileUrl:file.path,
+            owner:_id,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch(error) {
         return res.status(400).render("upload", {pageTitle:"Uploaded Video", errorMessage: error._message});
@@ -58,7 +72,18 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const {id} = req.params;
+    const {user:{_id}} = req.session;
+    const video = await Video.findById(id);
+    const user = await User.findById(_id);
+    if(!video){
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
+    user.videos.splice(user.videos.indexOf(id),1);
+    user.save();
     return res.redirect("/");
 };
